backend/requirements.txt
Flask==2.1.1
Flask-CORS==3.0.10
mysql-connector-python==8.0.29
